Add database.js file to connect to MongoDB.
Moved code that connected to MongoDB from server.js out to database.js to mimic the
old code base. Added nodemon as a developer dependency. Wasn't sure how to change
the start script to account for developer enviroment, so I just
replaced 'node' with 'nodemon'; this is fine for development, but the start script
needs to be changed when we move to production. Also wrote the challenges schema and
Challenge model. Added the mongo-santize package to replace the previous
database.santize function.