Make Challenges functioning.
We added the api.js file to interact with the database. We also added a route
to the api.js file. Now Challenges should be functioning. However, we still
need to account for some edge cases when adding new challenges, such as
ensuring that each challenge has a name and title as well as proper MIST code.
We should also use a better way to check a solution since currently we claim
correctness only when the code of the submission is identical to the challenge's
code; although it is unlikely for a submission to be correct and have different MIST code, it is still possible and we should find a way to deal with this edge case; perhaps include some way of simplifying MIST code or just checking if the
image is within a certain bound of correctness.