Translate the users database functions from MySQL to Mongoose/MongoDB
We translated the three functions changeAboutSection, getUser, getIDforUsername. Davin and Asya spent an hour on this. We still need to test these functions to make sure that they really work. We couldn't test changeAboutSection because profile.ejs is not working proper. We also added the about field to the users schema.