Fix profile page.
We should now be able to change the about section of the profile page and view profile pages of users. We also modified signupStrategy to replace the dateJoinedfield with the createdAt field. We fixed the routes in the userRouter and changed profile.ejs to be able to render the profile page of any user rather than just the one currently logged in.